Refactor url and view files
