+++ posts/urls.py
@@ -1,36 +1,50 @@
 from django.urls import path
 
-from users.views import *
-
-from . import views
-from .views import *
+from .views import (
+    add_comment_to_post,
+    check_link,
+    company_news,
+    create_space,
+    equity,
+    faq,
+    fixed_income,
+    like_post,
+    macro_economy,
+    my_account,
+    my_research,
+    post_detail,
+    post_edit,
+    post_list,
+    post_new,
+    post_remove,
+    post_share,
+    search,
+    space_detail,
+    space_list,
+)
 
 urlpatterns = [
-    path("home", views.post_list, name="post_list"),
-    path("spaces/", views.space_list, name="space_list"),
-    path("spaces/<int:pk>/", views.space_detail, name="space_detail"),
-    path("spaces/new/", views.create_space, name="create_space"),
-    path("<int:labels>/", views.post_list, name="post_list"),
-    path("post/<int:pk>/", views.post_detail, name="post_detail"),
-    path("post/new/", views.post_new, name="post_new"),
-    path("check_link/", views.check_link, name="check_link"),
-    path("post/<int:pk>/edit/", views.post_edit, name="post_edit"),
-    path("post/<int:pk>/remove/", views.post_remove, name="post_remove"),
-    path("post/<int:pk>/comment/", views.add_comment_to_post, name="add_comment_to_post"),
-    path("post/search/", views.search, name="search"),
-    path("myresearch/", views.my_research, name="my_research"),
-    path("myaccount/", views.my_account, name="my_account"),
-    path("faq/", views.faq, name="faq"),
-    path("macro_economy/", views.macro_economy, name="macro_economy"),
-    path("equity/", views.equity, name="equity"),
-    path("fixed_income/", views.fixed_income, name="fixed_income"),
-    path("company_news/", views.company_news, name="company_news"),
-    path("post/<int:pk>/share/", views.post_share, name="post_share"),
-    path("tag/<slug:tag_slug>/", views.post_list, name="post_list_by_tag"),
-    path("logout/", logout_request, name="logout"),
-    path("like/<int:pk>", views.like_post, name="like_post"),
-    path("users/", views.user_list, name="user_list"),
-    path("users/<username>/", views.user_detail, name="user_detail"),
-    path("users/follow", views.user_follow, name="user_follow"),
-    path("", views.post_list, name="post_list_"),
+    path("home", post_list, name="post_list"),
+    path("spaces/", space_list, name="space_list"),
+    path("spaces/new/", create_space, name="create_space"),
+    path("spaces/<int:pk>/", space_detail, name="space_detail"),
+    path("<int:labels>/", post_list, name="post_list"),
+    path("post/new/", post_new, name="post_new"),
+    path("post/<int:pk>/share/", post_share, name="post_share"),
+    path("post/<int:pk>/edit/", post_edit, name="post_edit"),
+    path("post/<int:pk>/remove/", post_remove, name="post_remove"),
+    path("post/<int:pk>/", post_detail, name="post_detail"),
+    path("post/<int:pk>/comment/", add_comment_to_post, name="add_comment_to_post"),
+    path("post/search/", search, name="search"),
+    path("myresearch/", my_research, name="my_research"),  # TODO: what is the difference between search
+    path("myaccount/", my_account, name="my_account"),
+    path("tag/<slug:tag_slug>/", post_list, name="post_list_by_tag"),
+    path("like/<int:pk>", like_post, name="like_post"),
+    path("check_link/", check_link, name="check_link"),
+    path("faq/", faq, name="faq"),
+    path("macro_economy/", macro_economy, name="macro_economy"),  # TODO: delete these  urls
+    path("equity/", equity, name="equity"),
+    path("fixed_income/", fixed_income, name="fixed_income"),
+    path("company_news/", company_news, name="company_news"),
+    path("", post_list, name="post_list_"),
 ]
