+++ users/urls.py
@@ -4,15 +4,15 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("login/", views.login_request, name="login"),
     path("register/", views.register_request, name="register"),
+    path("login/", views.login_request, name="login"),
     path("logout", views.logout_request, name="logout"),
     path(
         "password_change/",
         auth_views.PasswordChangeView.as_view(template_name="password_change_form.html"),
     ),
     path(
-        "password_change/done/",
+        "password_change/done/",  # TODO: check if done urls are necessary
         auth_views.PasswordChangeDoneView.as_view(template_name="password_change_done.html"),
     ),
     # reset password urls
@@ -32,5 +32,8 @@ urlpatterns = [
         "reset/done/",
         auth_views.PasswordResetCompleteView.as_view(template_name="password_reset_complete.html"),
     ),
-    path("edit/", views.edit, name="edit"),
+    path("edit/", views.edit, name="edit"),  # TODO: change name to more meaningful one
+    path("users/", views.user_list, name="user_list"),
+    path("users/<username>/", views.user_detail, name="user_detail"),
+    path("users/follow", views.user_follow, name="user_follow"),
 ]
