+++ users/views.py
@@ -1,11 +1,18 @@
 from django.contrib import messages
-from django.contrib.auth import authenticate, login
+from django.contrib.auth import authenticate, get_user_model, login
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import AuthenticationForm
 from django.contrib.auth.models import auth
-from django.shortcuts import redirect, render
+from django.db.models import Count
+from django.http import JsonResponse
+from django.shortcuts import get_object_or_404, redirect, render
+from django.views.decorators.http import require_POST
 
-from .forms import NewUserForm, ProfileEditForm, UserEditForm
-from .models import Profile
+from posts.models import Post
+from users.forms import NewUserForm, ProfileEditForm, UserEditForm
+from users.models import Contact, Profile
+
+User = get_user_model()
 
 
 def login_request(request):
@@ -42,12 +49,6 @@ def register_request(request):
         if form.is_valid():
             user = form.save()
             Profile.objects.create(user=user)
-            # a = Author()
-            # a.user_id = user.id
-            # a.email = user.email
-            # a.first_name = "test"
-            # a.last_name = "hehehe"
-            # a.save()
 
             login(request, user)
             messages.success(request, "Registration successful.")
@@ -90,4 +91,60 @@ def edit(request):
         request,
         "edit.html",
         {"user_form": user_form, "profile_form": profile_form},
+        # "posts:my_account.html",
+        # {"user_form": user_form, "profile_form": profile_form},  # TODO: which html should be used
+        # "posts:my_account.html",
     )
+
+
+@login_required
+def user_detail(request, username):
+    user = get_object_or_404(User, username=username, is_active=True)
+
+    all_posts = Post.objects.filter(author=user.id)
+    # most_commented_posts = Post.objects.filter(author_id=request.user.id).annotate(
+    #     total_comments=Count('comments')).order_by('-total_comments')[:3]
+    most_commented_posts = (
+        Post.objects.filter(author_id=user.id)
+        .annotate(total_comments=Count("comments"))
+        .order_by("-total_comments")[:4]
+    )
+    print(len(most_commented_posts))
+
+    return render(
+        request,
+        "posts/user/detail.html",
+        {
+            "section": "people",
+            "userToSee": user,
+            "id": user.id,
+            "latest_posts_db": all_posts[:4],
+            "count": all_posts.count(),
+            "most_commented_posts": most_commented_posts,
+        },
+    )
+
+
+@require_POST
+@login_required
+def user_follow(request):
+    user_id = request.POST.get("id")
+    action = request.POST.get("action")
+    if user_id and action:
+        try:
+            user = User.objects.get(id=user_id)
+            if action == "follow":
+                Contact.objects.get_or_create(user_from=request.user, user_to=user)
+
+            else:
+                Contact.objects.filter(user_from=request.user, user_to=user).delete()
+            return JsonResponse({"status": "ok"})
+        except User.DoesNotExist:
+            return JsonResponse({"status": "error"})
+    return JsonResponse({"status": "error"})
+
+
+@login_required
+def user_list(request):
+    users = User.objects.filter(is_active=True)
+    return render(request, "posts/user/list.html", {"section": "people", "users": users})
